e_shop/requirements.txt
alembic==1.4.1
asgiref==3.2.5
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.1
cryptography==2.8
Django==3.0.4
django-debug-toolbar==2.2
django-extensions==2.2.8
djangorestframework==3.11.0
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
paramiko==2.7.1
passlib==1.7.1
pgadmin4==4.17
pkg-resources==0.0.0
psutil==5.5.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.20
PyNaCl==1.3.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2018.9
simplejson==3.16.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.15
sqlparse==0.2.4
sshtunnel==0.1.5
Werkzeug==1.0.0
WTForms==2.2.1
